+++ zcash_keys/src/keys.rs
@@ -14,15 +14,15 @@ use zcash_protocol::consensus::NetworkConstants;
 
 #[cfg(feature = "transparent-inputs")]
 use {
+    ::transparent::keys::{IncomingViewingKey, NonHardenedChildIndex},
     core::convert::TryInto,
-    transparent::keys::{IncomingViewingKey, NonHardenedChildIndex},
 };
 
 #[cfg(all(
     feature = "transparent-inputs",
     any(test, feature = "test-dependencies")
 ))]
-use transparent::address::TransparentAddress;
+use ::transparent::address::TransparentAddress;
 
 #[cfg(feature = "unstable")]
 use {
@@ -78,6 +78,9 @@ pub mod sapling {
         )
     }
 }
+
+#[cfg(feature = "transparent-key-encoding")]
+pub mod transparent;
 
 #[cfg(feature = "transparent-inputs")]
 fn to_transparent_child_index(j: DiversifierIndex) -> Option<NonHardenedChildIndex> {
@@ -207,7 +210,7 @@ impl Era {
 #[derive(Clone, Debug)]
 pub struct UnifiedSpendingKey {
     #[cfg(feature = "transparent-inputs")]
-    transparent: transparent::keys::AccountPrivKey,
+    transparent: ::transparent::keys::AccountPrivKey,
     #[cfg(feature = "sapling")]
     sapling: sapling::ExtendedSpendingKey,
     #[cfg(feature = "orchard")]
@@ -226,7 +229,7 @@ impl UnifiedSpendingKey {
 
         UnifiedSpendingKey::from_checked_parts(
             #[cfg(feature = "transparent-inputs")]
-            transparent::keys::AccountPrivKey::from_seed(_params, seed, _account)
+            ::transparent::keys::AccountPrivKey::from_seed(_params, seed, _account)
                 .map_err(DerivationError::Transparent)?,
             #[cfg(feature = "sapling")]
             sapling::spending_key(seed, _params.coin_type(), _account),
@@ -239,7 +242,7 @@ impl UnifiedSpendingKey {
     /// Construct a USK from its constituent parts, after verifying that UIVK derivation can
     /// succeed.
     fn from_checked_parts(
-        #[cfg(feature = "transparent-inputs")] transparent: transparent::keys::AccountPrivKey,
+        #[cfg(feature = "transparent-inputs")] transparent: ::transparent::keys::AccountPrivKey,
         #[cfg(feature = "sapling")] sapling: sapling::ExtendedSpendingKey,
         #[cfg(feature = "orchard")] orchard: orchard::keys::SpendingKey,
     ) -> Result<UnifiedSpendingKey, DerivationError> {
@@ -273,7 +276,7 @@ impl UnifiedSpendingKey {
     /// Returns the transparent component of the unified key at the
     /// BIP44 path `m/44'/<coin_type>'/<account>'`.
     #[cfg(feature = "transparent-inputs")]
-    pub fn transparent(&self) -> &transparent::keys::AccountPrivKey {
+    pub fn transparent(&self) -> &::transparent::keys::AccountPrivKey {
         &self.transparent
     }
 
@@ -418,7 +421,7 @@ impl UnifiedSpendingKey {
                     #[cfg(feature = "transparent-inputs")]
                     {
                         transparent = Some(
-                            transparent::keys::AccountPrivKey::from_bytes(&key)
+                            ::transparent::keys::AccountPrivKey::from_bytes(&key)
                                 .ok_or(DecodingError::KeyDataInvalid(Typecode::P2pkh))?,
                         );
                     }
@@ -711,7 +714,7 @@ impl From<bip32::Error> for DerivationError {
 #[derive(Clone, Debug)]
 pub struct UnifiedFullViewingKey {
     #[cfg(feature = "transparent-inputs")]
-    transparent: Option<transparent::keys::AccountPubKey>,
+    transparent: Option<::transparent::keys::AccountPubKey>,
     #[cfg(feature = "sapling")]
     sapling: Option<sapling::DiversifiableFullViewingKey>,
     #[cfg(feature = "orchard")]
@@ -728,7 +731,7 @@ impl UnifiedFullViewingKey {
     #[cfg(any(test, feature = "test-dependencies"))]
     pub fn new(
         #[cfg(feature = "transparent-inputs")] transparent: Option<
-            transparent::keys::AccountPubKey,
+            ::transparent::keys::AccountPubKey,
         >,
         #[cfg(feature = "sapling")] sapling: Option<sapling::DiversifiableFullViewingKey>,
         #[cfg(feature = "orchard")] orchard: Option<orchard::keys::FullViewingKey>,
@@ -785,7 +788,7 @@ impl UnifiedFullViewingKey {
     /// succeed.
     fn from_checked_parts(
         #[cfg(feature = "transparent-inputs")] transparent: Option<
-            transparent::keys::AccountPubKey,
+            ::transparent::keys::AccountPubKey,
         >,
         #[cfg(feature = "sapling")] sapling: Option<sapling::DiversifiableFullViewingKey>,
         #[cfg(feature = "orchard")] orchard: Option<orchard::keys::FullViewingKey>,
@@ -872,7 +875,7 @@ impl UnifiedFullViewingKey {
                     data.to_vec(),
                 ))),
                 #[cfg(feature = "transparent-inputs")]
-                unified::Fvk::P2pkh(data) => transparent::keys::AccountPubKey::deserialize(data)
+                unified::Fvk::P2pkh(data) => ::transparent::keys::AccountPubKey::deserialize(data)
                     .map_err(|_| DecodingError::KeyDataInvalid(Typecode::P2pkh))
                     .map(|tfvk| {
                         transparent = Some(tfvk);
@@ -958,7 +961,7 @@ impl UnifiedFullViewingKey {
     /// Returns the transparent component of the unified key at the
     /// BIP44 path `m/44'/<coin_type>'/<account>'`.
     #[cfg(feature = "transparent-inputs")]
-    pub fn transparent(&self) -> Option<&transparent::keys::AccountPubKey> {
+    pub fn transparent(&self) -> Option<&::transparent::keys::AccountPubKey> {
         self.transparent.as_ref()
     }
 
@@ -1021,7 +1024,7 @@ impl UnifiedFullViewingKey {
 #[derive(Clone, Debug)]
 pub struct UnifiedIncomingViewingKey {
     #[cfg(feature = "transparent-inputs")]
-    transparent: Option<transparent::keys::ExternalIvk>,
+    transparent: Option<::transparent::keys::ExternalIvk>,
     #[cfg(feature = "sapling")]
     sapling: Option<::sapling::zip32::IncomingViewingKey>,
     #[cfg(feature = "orchard")]
@@ -1038,7 +1041,9 @@ impl UnifiedIncomingViewingKey {
     /// be used instead.
     #[cfg(any(test, feature = "test-dependencies"))]
     pub fn new(
-        #[cfg(feature = "transparent-inputs")] transparent: Option<transparent::keys::ExternalIvk>,
+        #[cfg(feature = "transparent-inputs")] transparent: Option<
+            ::transparent::keys::ExternalIvk,
+        >,
         #[cfg(feature = "sapling")] sapling: Option<::sapling::zip32::IncomingViewingKey>,
         #[cfg(feature = "orchard")] orchard: Option<orchard::keys::IncomingViewingKey>,
         // TODO: Implement construction of UIVKs with metadata items.
@@ -1115,7 +1120,7 @@ impl UnifiedIncomingViewingKey {
                     #[cfg(feature = "transparent-inputs")]
                     {
                         transparent = Some(
-                            transparent::keys::ExternalIvk::deserialize(data)
+                            ::transparent::keys::ExternalIvk::deserialize(data)
                                 .map_err(|_| DecodingError::KeyDataInvalid(Typecode::P2pkh))?,
                         );
                     }
@@ -1191,7 +1196,7 @@ impl UnifiedIncomingViewingKey {
 
     /// Returns the Transparent external IVK, if present.
     #[cfg(feature = "transparent-inputs")]
-    pub fn transparent(&self) -> &Option<transparent::keys::ExternalIvk> {
+    pub fn transparent(&self) -> &Option<::transparent::keys::ExternalIvk> {
         &self.transparent
     }
 
@@ -1506,9 +1511,9 @@ mod tests {
     #[cfg(feature = "transparent-inputs")]
     use {
         crate::{address::Address, encoding::AddressCodec},
+        ::transparent::keys::{AccountPrivKey, IncomingViewingKey},
         alloc::string::ToString,
         alloc::vec::Vec,
-        transparent::keys::{AccountPrivKey, IncomingViewingKey},
         zcash_address::test_vectors,
         zip32::DiversifierIndex,
     };
@@ -1535,7 +1540,7 @@ mod tests {
     #[cfg(feature = "transparent-inputs")]
     #[test]
     fn pk_to_taddr() {
-        use transparent::keys::NonHardenedChildIndex;
+        use ::transparent::keys::NonHardenedChildIndex;
 
         let taddr = AccountPrivKey::from_seed(&MAIN_NETWORK, &seed(), AccountId::ZERO)
             .unwrap()